Reschedule jobs on failure
